Refactor Item and Checklist models to use Integer for IDs, update ItemController and ItemRepository accordingly

+++ src/main/java/com/teofilus/todoapp/Controller/ItemController.java
@@ -29,7 +29,7 @@ public class ItemController {
 
     @GetMapping
     public List<Item> getItems(@PathVariable Integer checklistId) {
-        return itemService.getItemsByChecklistId(checklistId.longValue());
+        return itemService.getItemsByChecklistId(checklistId);
     }
 
     @PostMapping
